+++ app/services/manufacturer.py
@@ -1,8 +1,7 @@
-from app.configs.exception.exception import NotFoundError
+from app.configs.exception.exception import CannotDelete, NotFoundError
 from app.repositories import ManufacturerRepository
 from app.models import Manufacturer
-from app.schemes import ManufacturerCreateScheme, ManufacturerUpdateScheme
-from app.schemes.generic_params import GenericFilterParams
+from app.schemes import ManufacturerCreateScheme, ManufacturerUpdateScheme, ManufacturerParamsScheme
 
 
 class ManufacturerService:
@@ -10,7 +9,7 @@ class ManufacturerService:
     def __init__(self, manufacturer_repository: ManufacturerRepository) -> None:
         self._repository: ManufacturerRepository = manufacturer_repository
 
-    def get_manufacturers(self, params: GenericFilterParams) -> list[Manufacturer]:
+    def get_manufacturers(self, params: ManufacturerParamsScheme) -> list[Manufacturer]:
         return self._repository.get_paginated_list(params)
 
     def get_manufacturer_by_id(self, manufacturer_id: int) -> Manufacturer:
@@ -31,4 +30,6 @@ class ManufacturerService:
 
     def delete_manufacturer_by_id(self, manufacturer_id: int) -> None:
         manufacturer = self.get_manufacturer_by_id(manufacturer_id)
+        if manufacturer.models:
+            raise CannotDelete("Cannot delete manufacturer with models assigned.")
         return self._repository.delete(manufacturer)
